+++ Makefile.am
@@ -122,7 +122,7 @@ nodist_src_libfabric_la_SOURCES =
 src_libfabric_la_SOURCES =			\
 	include/fi.h				\
 	include/ofi_abi.h			\
-	include/fi_atom.h			\
+	include/ofi_atom.h			\
 	include/fi_enosys.h			\
 	include/ofi_file.h			\
 	include/fi_indexer.h			\
