+++ include/ofi_mr.h
@@ -41,7 +41,7 @@
 #include <stdbool.h>
 
 #include <fi.h>
-#include <fi_atom.h>
+#include <ofi_atom.h>
 #include <fi_lock.h>
 #include <fi_list.h>
 #include <rbtree.h>
