+++ libfabric.vcxproj
@@ -566,7 +566,7 @@
     <ClInclude Include="include\fasthash.h" />
     <ClInclude Include="include\fi.h" />
     <ClInclude Include="include\ofi_abi.h" />
-    <ClInclude Include="include\fi_atom.h" />
+    <ClInclude Include="include\ofi_atom.h" />
     <ClInclude Include="include\ofi_atomic.h" />
     <ClInclude Include="include\ofi_mr.h" />
     <ClInclude Include="include\ofi_net.h" />
